Add circular dependency test
