+++ frontend/src/components/types.ts
@@ -18,3 +18,25 @@ export type IpTripletSelectorValidationType = {
   message?: string;
   triplets: IpTripletProps['validated'][];
 };
+
+export type K8SStateContextData = {
+  username: string;
+  usernameValidation?: string; // just a message or empty
+  handleSetUsername: (newVal: string) => void;
+
+  password: string;
+  passwordValidation?: string;
+  handleSetPassword: (newVal: string) => void;
+
+  apiaddr: string; // 12 characters
+  apiaddrValidation: IpTripletSelectorValidationType;
+  handleSetApiaddr: (newApiaddr: string) => void;
+
+  ingressIp: string; // 12 characters
+  ingressIpValidation: IpTripletSelectorValidationType;
+  handleSetIngressIp: (newIp: string) => void;
+
+  domain: string;
+  handleSetDomain: (newDomain: string) => void;
+  domainValidation?: string;
+};
