+++ frontend/src/resources/utils.ts
@@ -1,7 +1,3 @@
-import React from 'react';
-import { getResource } from './resource-request';
-import { Route } from './Route';
-
 export const pathJoin = (...args: string[]) => {
   return args
     .map((part, i) => {
@@ -13,28 +9,4 @@ export const pathJoin = (...args: string[]) => {
     })
     .filter((x) => x.length)
     .join('/');
-};
-
-export const useConsoleUrl = () => {
-  const [consoleUrl, setConsoleUrl] = React.useState<string>();
-  React.useEffect(() => {
-    const doItAsync = async () => {
-      const route = await getResource<Route>({
-        apiVersion: 'route.openshift.io/v1',
-        kind: 'Route',
-        metadata: {
-          name: 'console',
-          namespace: 'openshift-console',
-        },
-      }).promise;
-
-      if (route.spec?.host) {
-        const result = `https://${route.spec.host}`;
-        console.log('OCP console URL read: ', result);
-        setConsoleUrl(result);
-      }
-    };
-    doItAsync();
-  }, []);
-  return consoleUrl;
 };
